Add GroupAdjacentBy()
this method helps to group a collection of dates into sequences of
contiguous dates.  this is useful for querying web apis, when we
want to consolidate web calls over date ranges.  instead of three
queries for 2018.1.1, 2018.1.2 and 2018.1.4, we can consolidate
the first two queries into one, given adjacent groupings of dates